Tape tests are all the same.
