+++ src/tape/vec_tape.rs
@@ -84,61 +84,4 @@ impl ops::DerefMut for VecTape {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use tape::Tape;
-
-    #[test]
-    fn new() {
-        let _ = VecTape::default();
-    }
-
-    #[test]
-    fn deref() {
-        let tape = VecTape::default();
-        assert_eq!(*tape, 0);
-    }
-
-    #[test]
-    fn deref_mut() {
-        let mut tape = VecTape::default();
-        tape.inc_val().unwrap();
-        *tape = 20;
-        assert_eq!(*tape, 20);
-    }
-
-    #[test]
-    fn inc_val() {
-        let mut tape = VecTape::default();
-        *tape = 20;
-        tape.inc_val().unwrap();
-        assert_eq!(*tape, 21);
-    }
-
-    #[test]
-    fn dec_val() {
-        let mut tape = VecTape::default();
-        *tape = 20;
-        tape.dec_val().unwrap();
-        assert_eq!(*tape, 19);
-    }
-
-    #[test]
-    fn inc_ptr() {
-        let mut tape = VecTape::default();
-        *tape = 20;
-        tape.inc_ptr().unwrap();
-        assert_eq!(*tape, 0);
-    }
-
-    #[test]
-    fn dec_ptr() {
-        let mut tape = VecTape::default();
-        *tape = 20;
-        tape.inc_ptr().unwrap();
-        assert_eq!(*tape, 0);
-        tape.dec_ptr().unwrap();
-        assert_eq!(*tape, 20);
-    }
-}
+tape_tests!(VecTape);
